Move modal to seperate component
